+++ src/main/java/com/raoulvdberge/refinedstorage/container/ContainerCraftingMonitor.java
@@ -1,28 +1,33 @@
 package com.raoulvdberge.refinedstorage.container;
 
+import com.raoulvdberge.refinedstorage.gui.IResizableDisplay;
 import com.raoulvdberge.refinedstorage.tile.craftingmonitor.ICraftingMonitor;
 import com.raoulvdberge.refinedstorage.tile.craftingmonitor.TileCraftingMonitor;
 import com.raoulvdberge.refinedstorage.tile.craftingmonitor.WirelessCraftingMonitor;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
-import net.minecraftforge.items.SlotItemHandler;
 
 import javax.annotation.Nullable;
 
 public class ContainerCraftingMonitor extends ContainerBase {
     private ICraftingMonitor craftingMonitor;
+    private IResizableDisplay resizableDisplay;
 
-    public ContainerCraftingMonitor(ICraftingMonitor craftingMonitor, @Nullable TileCraftingMonitor craftingMonitorTile, EntityPlayer player) {
+    public ContainerCraftingMonitor(ICraftingMonitor craftingMonitor, @Nullable TileCraftingMonitor craftingMonitorTile, EntityPlayer player, IResizableDisplay resizableDisplay) {
         super(craftingMonitorTile, player);
 
         this.craftingMonitor = craftingMonitor;
+        this.resizableDisplay = resizableDisplay;
 
-        for (int i = 0; i < 4; ++i) {
-            addSlotToContainer(new SlotItemHandler(craftingMonitor.getFilter(), i, 187, 6 + (18 * i)));
-        }
+        initSlots();
+    }
 
-        addPlayerInventory(8, 148);
+    public void initSlots() {
+        this.inventorySlots.clear();
+        this.inventoryItemStacks.clear();
+
+        addPlayerInventory(8, resizableDisplay.getYPlayerInventory());
     }
 
     @Override