+++ src/main/java/com/raoulvdberge/refinedstorage/tile/craftingmonitor/WirelessCraftingMonitor.java
@@ -1,23 +1,21 @@
 package com.raoulvdberge.refinedstorage.tile.craftingmonitor;
 
+import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.api.autocrafting.task.ICraftingTask;
 import com.raoulvdberge.refinedstorage.api.network.INetwork;
-import com.raoulvdberge.refinedstorage.api.util.IFilter;
-import com.raoulvdberge.refinedstorage.inventory.ItemHandlerBase;
-import com.raoulvdberge.refinedstorage.inventory.ItemHandlerFilter;
+import com.raoulvdberge.refinedstorage.gui.GuiBase;
+import com.raoulvdberge.refinedstorage.gui.GuiCraftingMonitor;
 import com.raoulvdberge.refinedstorage.item.ItemWirelessCraftingMonitor;
+import com.raoulvdberge.refinedstorage.network.MessageWirelessCraftingMonitorSize;
 import com.raoulvdberge.refinedstorage.tile.data.TileDataParameter;
-import com.raoulvdberge.refinedstorage.util.StackUtils;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.common.DimensionManager;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -26,35 +24,13 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
 
     private int networkDimension;
     private BlockPos network;
-
-    private List<IFilter> filters = new ArrayList<>();
-    private ItemHandlerFilter filter = new ItemHandlerFilter(filters, new ArrayList<>(), null) {
-        @Override
-        protected void onContentsChanged(int slot) {
-            super.onContentsChanged(slot);
-
-            if (!stack.hasTagCompound()) {
-                stack.setTagCompound(new NBTTagCompound());
-            }
-
-            StackUtils.writeItems(this, 0, stack.getTagCompound());
-
-            INetwork network = getNetwork();
-
-            if (network != null) {
-                network.getCraftingManager().sendCraftingMonitorUpdate();
-            }
-        }
-    };
+    private int size;
 
     public WirelessCraftingMonitor(int networkDimension, ItemStack stack) {
         this.stack = stack;
         this.networkDimension = networkDimension;
         this.network = new BlockPos(ItemWirelessCraftingMonitor.getX(stack), ItemWirelessCraftingMonitor.getY(stack), ItemWirelessCraftingMonitor.getZ(stack));
-
-        if (stack.hasTagCompound()) {
-            StackUtils.readItems(filter, 0, stack.getTagCompound());
-        }
+        this.size = ItemWirelessCraftingMonitor.getSize(stack);
     }
 
     @Override
@@ -93,13 +69,17 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
     }
 
     @Override
-    public List<IFilter> getFilters() {
-        return filters;
+    public int getSize() {
+        return size;
     }
 
     @Override
-    public ItemHandlerBase getFilter() {
-        return filter;
+    public void onSizeChanged(int size) {
+        this.size = size;
+
+        GuiBase.executeLater(GuiCraftingMonitor.class, GuiBase::initGui);
+
+        RS.INSTANCE.network.sendToServer(new MessageWirelessCraftingMonitorSize(size));
     }
 
     private INetwork getNetwork() {