+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed bug where pattern was recipe pattern was creatable when there was no recipe output (raoulvdberge)
 - If an Interface is configured to expose the entire network storage (by configuring no export slots), it will no longer expose the entire RS storage, due to performance issues (raoulvdberge)
 - The Portable Grid no longer exposes a inventory for crossmod interaction, due to performance issues (raoulvdberge)
+- The Crafting Monitor is now resizable and its size can be configured (stretched, small, medium, large) (raoulvdberge)
 - Removed the Wrench (raoulvdberge)
 - Updated Russian translation (kellixon)